chore: break out buildPackage into separate files
